Update some attributes in models and Add findById methods for ClienteService and ProutoService

+++ implementacao/src/main/java/com/ithappens/interview/dtos/FilialProdutoDTO.java
@@ -20,7 +20,7 @@ public class FilialProdutoDTO {
 
     private String codigoDeBarras;
 
-    private Long sequencial;
+    private Integer sequencial;
 
     private Double custo;
